Enable listings map button
