+++ containers/views/Listings/Results/styles.js
@@ -1,0 +1,13 @@
+import {StyleSheet} from 'react-native'
+
+export default StyleSheet.create({
+  container: {
+    flex: 1,
+    position: 'relative'
+  },
+  mapButton: {
+    position: 'absolute',
+    right: 20,
+    bottom: 20
+  }
+})
